[SU.2] Fix some -Weverything

+++ src/su/cs-tools.c
@@ -36,7 +36,7 @@ su_cs_cmp(char const *cp1, char const *cp2){
    ASSERT_NYD_EXEC(cp2 != NIL, rv = 1);
 
    for(;;){
-      u8 c1, c2;
+      s32 c1, c2;
 
       c1 = *cp1++;
       c2 = *cp2++;
@@ -56,7 +56,7 @@ su_cs_cmp_n(char const *cp1, char const *cp2, uz n){
    ASSERT_NYD_EXEC(cp2 != NIL, rv = 1);
 
    for(rv = 0; n != 0; --n){
-      u8 c1, c2;
+      s32 c1, c2;
 
       c1 = *cp1++;
       c2 = *cp2++;
